requirements.txt
Flask
Flask-Bcrypt
Flask-Session
pymongo
python-dotenv